working with .bashrc
Although this recipe works with ~/.bash_profile, I couldn't for the life of me get the serverspec tests to pass.  It turns out serverspec does not create a login shell when running its tests. Using .bashrc works, though, so that's what I've decided to go with.  You could easily replace bashrc with bash_profile for practical reasons, but the spec would fail.